requirements.txt
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
django-cors-headers==3.13.0
django-environ==0.8.1
celery[redis]==5.0.5
django-celery-beat==2.2.1
psycopg2==2.9.3
django-environ==0.8.1
PyGithub==1.55
python-gitlab==3.4.0
slack-sdk==3.16.2
atlassian-python-api==3.23.0
gevent==21.12.0
boto3==1.26.32
pyyaml==6.0
uritemplate==4.1.1